fix: preserve expected pins across the editor toggle
